(electron) Redirect typescript output
