+++ builderjs/src/index.test.tmp
@@ -1,0 +1,8 @@
+import { GetLocalModules } from "./index";
+//const { GetLocalModules } = require(".");
+
+test('Placeholder', () => {
+    expect(true).toBe(true);
+});
+
+console.log(GetLocalModules("../../snakeshack"));
