+++ builderjs/src/index.ts
@@ -1,8 +1,12 @@
-"use strict";
-const yaml = require("js-yaml");
-const path = require("path");
-const fs = require("fs");
-const GetModulesList = (url) => {
+import yaml from "js-yaml";
+import path from "path";
+import fs from "fs";
+
+//const yaml = require("js-yaml");
+//const path = require("path");
+//const fs = require("fs");
+
+const GetModulesList = (url: Record<string, any>) => {
   console.log("GetModulesList", url);
   switch (url["type"]) {
     case "github":
@@ -13,25 +17,31 @@ const GetModulesList = (url) => {
       throw new Error("Invalid url type.");
   }
 };
-const GetFolders = (root_folder) =>
+
+const GetFolders = (root_folder: string): Array<string> =>
   fs
     .readdirSync(root_folder, { withFileTypes: true })
-    .filter((f) => f.isDirectory())
-    .map((f) => f.name);
-const GetLocalModules = (root_folder) => {
+    .filter((f: any) => f.isDirectory())
+    .map((f: any) => f.name);
+
+const GetLocalModules = (root_folder: string): Array<Record<string, any>> => {
   // static return for now
   const path_base = path.join(path.resolve(root_folder), "workflows");
+
   // Get list of local filesystem directories in path
   const orgs = GetFolders(path_base);
+
   // First-level (organisation) listing
   const modules = [];
   for (const org of orgs) {
     const org_path = path.join(path_base, org);
     const module_types = GetFolders(org_path).reverse();
+
     // Second-level (module type) listing
     for (const module_type of module_types) {
       const module_type_path = path.join(org_path, module_type);
       const workflows = GetFolders(module_type_path).sort();
+
       // Third-level (module/workflow) listing
       for (const workflow of workflows) {
         const url_workflow = path.join(
@@ -48,15 +58,19 @@ const GetLocalModules = (root_folder) => {
           workflow,
           "config/config.yaml"
         );
+
         let params = {};
         try {
-          params = yaml.load(fs.readFileSync(config_file, "utf8"));
+          params = yaml.load(fs.readFileSync(config_file, "utf8")) as Record<
+            string,
+            any
+          >;
         } catch (err) {
           console.log("No (or invalid YAML) config file found.");
         }
         modules.push({
           name: "(" + org + ") " + FormatName(workflow),
-          type: module_type.slice(0, -1),
+          type: module_type.slice(0, -1), // remove plural
           config: {
             url: url_workflow,
             params: params,
@@ -67,7 +81,11 @@ const GetLocalModules = (root_folder) => {
   }
   return modules;
 };
-const GetRemoteModulesGithub = (repo, listing_type) => {
+
+const GetRemoteModulesGithub = (
+  repo: string,
+  listing_type: string
+): Array<Record<string, any>> => {
   switch (listing_type) {
     case "DirectoryListing":
       return GetRemoteModulesGithubDirectoryListing(repo);
@@ -77,7 +95,10 @@ const GetRemoteModulesGithub = (repo, listing_type) => {
       throw new Error("Invalid Github listing type.");
   }
 };
-const GetRemoteModulesGithubDirectoryListing = (repo) => {
+
+const GetRemoteModulesGithubDirectoryListing = (
+  repo: string
+): Array<Record<string, any>> => {
   // static return for now
   const modules = [
     {
@@ -99,7 +120,10 @@ const GetRemoteModulesGithubDirectoryListing = (repo) => {
   ];
   return modules;
 };
-const GetRemoteModulesGithubBranchListing = (repo) => {
+
+const GetRemoteModulesGithubBranchListing = (
+  repo: string
+): Array<Record<string, any>> => {
   // static return for now
   const modules = [
     {
@@ -121,8 +145,11 @@ const GetRemoteModulesGithubBranchListing = (repo) => {
   ];
   return modules;
 };
-const FormatName = (name) => {
+
+const FormatName = (name: string): string => {
   return name;
 };
+
 // CommonJS export methods (not ES6)
-module.exports = { GetLocalModules, GetModulesList };
+//module.exports = { GetLocalModules, GetModulesList };
+export { GetLocalModules, GetModulesList };
