--- builderjs/test_index.js
@@ -1,3 +1,0 @@
-const { GetLocalModules } = require("./index.js");
-
-console.log(GetLocalModules("../../snakeshack"));
