+++ electron-app/.gitignore
@@ -1,3 +1,4 @@
 .DS_Store
 out
 node_modules
+dist
